+++ internal/provider/lldpd_config_resource.go
@@ -49,6 +49,7 @@ func (r *lldpdConfigResource) Configure(_ context.Context, req resource.Configur
 func (r *lldpdConfigResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		Description: "Lldpd config.",
+		Version:     1,
 		Attributes: map[string]schema.Attribute{
 			"id":                computedIDAttribute(),
 			"managed":           managedAttribute(),