+++ internal/provider/system_resource.go
@@ -50,6 +50,7 @@ func (r *systemResource) Configure(_ context.Context, req resource.ConfigureRequ
 func (r *systemResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		Description: "System settings.",
+		Version:     1,
 		Attributes: map[string]schema.Attribute{
 			"id":           computedIDAttribute(),
 			"managed":      managedAttribute(),